components ausgelagert
